pyproject.toml
[project]
name = "resturant-order"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "groq>=1.1.2",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
